Update Convert Fahrenheit to Celsius.js (#361)
* Update Convert Fahrenheit to Celsius.js

Modified parser to convert all Fahrenheit temperatures in the input string to Celsius instead of only the first temp in the string.

E.x. 

"it went from 45F to 25F overnight and then back to 55F the next day"

Produces
45°F is 7.22 degrees in sane units (Celsius).
25°F is -3.89 degrees in sane units (Celsius).
55°F is 12.78 degrees in sane units (Celsius).

* Update Convert Fahrenheit to Celsius.js

Update F to C temperature parser to convert all temperatures in the input string, instead of just 1.

Updated parser for ECMAScript 2021 compatibility.